+++ backend/test/handlers/WidgetHandler.test.ts
@@ -1,7 +1,7 @@
 // @ts-nocheck File not migrated fully to TS
 import mkdirp from 'mkdirp';
 import { getResourcePath } from 'utils';
-import { listWidgets } from 'handler/WidgetHandler';
+import { listWidgets } from 'handler/WidgetsHandler';
 
 jest.mock('handler/ManagerHandler');
